dom/imptests/webapps/DOMCore/tests/approved/test_Range-extractContents.html

<!doctype html>
<title>Range.extractContents() tests</title>
<link rel=author title="Aryeh Gregor" href=ayg@aryeh.name>
<p>To debug test failures, look at the URL of the hidden iframes corresponding
to the test at the end of the document, load those individually, and use your
browser's debugging tools.
<div id=log></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=common.js></script>
<script>
"use strict";

testDiv.parentNode.removeChild(testDiv);

function testExtractContents(i) {
	// We only load the expected once the actual is finished, to avoid race
	// conditions.  First set the onload to null and the src to the empty
	// string, so that when we set the src for real we're not just changing the
	// hash -- otherwise the load event won't fire.  We never do anything after
	// a src set except in an onload handler, to avoid race conditions.
	var actual = actualIframes[i];
	actual.id = i + "-actual";
	actual.onload = function() { onLoadActual(i); };
	actual.src = "Range-test-iframe.html#" + testRanges[i];
}

function onLoadActual(i) {
	var expected = expectedIframes[i];
	expected.id = i + "-expected";
	expected.onload = function() { onLoadExpected(i); };
	expected.src = "Range-test-iframe.html#" + testRanges[i];
}

function onLoadExpected(i) {
	var actualRange = actualIframes[i].contentWindow.testRange;
	var expectedRange = expectedIframes[i].contentWindow.testRange;
	var actualFrag, expectedFrag;
	var actualRoots, expectedRoots;

	domTests[i].step(function() {
		assert_equals(actualIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for actual extractContents()");
		assert_equals(expectedIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for simulated extractContents()");
		assert_equals(typeof actualRange, "object",
			"typeof Range produced in actual iframe");
		assert_equals(typeof expectedRange, "object",
			"typeof Range produced in expected iframe");

		// Just to be pedantic, we'll test not only that the tree we're
		// modifying is the same in expected vs. actual, but also that all the
		// nodes originally in it were the same.  Typically some nodes will
		// become detached when the algorithm is run, but they still exist and
		// references can still be kept to them, so they should also remain the
		// same.
		//
		// We initialize the list to all nodes, and later on remove all the
		// ones which still have parents, since the parents will presumably be
		// tested for isEqualNode() and checking the children would be
		// redundant.
		var actualAllNodes = [];
		var node = furthestAncestor(actualRange.startContainer);
		do {
			actualAllNodes.push(node);
		} while (node = nextNode(node));

		var expectedAllNodes = [];
		var node = furthestAncestor(expectedRange.startContainer);
		do {
			expectedAllNodes.push(node);
		} while (node = nextNode(node));

		expectedFrag = myExtractContents(expectedRange);
		if (typeof expectedFrag == "string") {
			assert_throws(expectedFrag, function() {
				actualRange.extractContents();
			});
		} else {
			actualFrag = actualRange.extractContents();
		}

		actualRoots = [];
		for (var j = 0; j < actualAllNodes.length; j++) {
			if (!actualAllNodes[j].parentNode) {
				actualRoots.push(actualAllNodes[j]);
			}
		}

		expectedRoots = [];
		for (var j = 0; j < expectedAllNodes.length; j++) {
			if (!expectedAllNodes[j].parentNode) {
				expectedRoots.push(expectedAllNodes[j]);
			}
		}

		for (var j = 0; j < actualRoots.length; j++) {
			assertNodesEqual(actualRoots[j], expectedRoots[j], j ? "detached node #" + j : "tree root");

			if (j == 0) {
				// Clearly something is wrong if the node lists are different
				// lengths.  We want to report this only after we've already
				// checked the main tree for equality, though, so it doesn't
				// mask more interesting errors.
				assert_equals(actualRoots.length, expectedRoots.length,
					"Actual and expected DOMs were broken up into a different number of pieces by extractContents() (this probably means you created or detached nodes when you weren't supposed to)");
			}
		}
	});
	domTests[i].done();

	positionTests[i].step(function() {
		assert_equals(actualIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for actual extractContents()");
		assert_equals(expectedIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for simulated extractContents()");
		assert_equals(typeof actualRange, "object",
			"typeof Range produced in actual iframe");
		assert_equals(typeof expectedRange, "object",
			"typeof Range produced in expected iframe");

		assert_true(actualRoots[0].isEqualNode(expectedRoots[0]),
			"The resulting DOMs were not equal, so comparing positions makes no sense");

		if (typeof expectedFrag == "string") {
			// It's no longer true that, e.g., startContainer and endContainer
			// must always be the same
			return;
		}
		assert_equals(actualRange.startContainer, actualRange.endContainer,
			"startContainer and endContainer must always be the same after extractContents()");
		assert_equals(actualRange.startOffset, actualRange.endOffset,
			"startOffset and endOffset must always be the same after extractContents()");
		assert_equals(expectedRange.startContainer, expectedRange.endContainer,
			"Test bug!  Expected startContainer and endContainer must always be the same after extractContents()");
		assert_equals(expectedRange.startOffset, expectedRange.endOffset,
			"Test bug!  Expected startOffset and endOffset must always be the same after extractContents()");

		assert_equals(actualRange.startOffset, expectedRange.startOffset,
			"Unexpected startOffset after extractContents()");
		// How do we decide that the two nodes are equal, since they're in
		// different trees?  Since the DOMs are the same, it's enough to check
		// that the index in the parent is the same all the way up the tree.
		// But we can first cheat by just checking they're actually equal.
		assert_true(actualRange.startContainer.isEqualNode(expectedRange.startContainer),
			"Unexpected startContainer after extractContents(), expected " +
			expectedRange.startContainer.nodeName.toLowerCase() + " but got " +
			actualRange.startContainer.nodeName.toLowerCase());
		var currentActual = actualRange.startContainer;
		var currentExpected = expectedRange.startContainer;
		var actual = "";
		var expected = "";
		while (currentActual && currentExpected) {
			actual = indexOf(currentActual) + "-" + actual;
			expected = indexOf(currentExpected) + "-" + expected;

			currentActual = currentActual.parentNode;
			currentExpected = currentExpected.parentNode;
		}
		actual = actual.substr(0, actual.length - 1);
		expected = expected.substr(0, expected.length - 1);
		assert_equals(actual, expected,
			"startContainer superficially looks right but is actually the wrong node if you trace back its index in all its ancestors (I'm surprised this actually happened");
	});
	positionTests[i].done();

	fragTests[i].step(function() {
		assert_equals(actualIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for actual extractContents()");
		assert_equals(expectedIframes[i].contentWindow.unexpectedException, null,
			"Unexpected exception thrown when setting up Range for simulated extractContents()");
		assert_equals(typeof actualRange, "object",
			"typeof Range produced in actual iframe");
		assert_equals(typeof expectedRange, "object",
			"typeof Range produced in expected iframe");

		if (typeof expectedFrag == "string") {
			// Comparing makes no sense
			return;
		}
		assertNodesEqual(actualFrag, expectedFrag,
			"returned fragment");
	});
	fragTests[i].done();
}

setup({timeout: 60000});

// First test a detached Range, synchronously
test(function() {
	var range = document.createRange();
	range.detach();
	assert_throws("INVALID_STATE_ERR", function() {
		range.extractContents();
	}, "Detached Range must throw INVALID_STATE_ERR on .extractContents()");
}, "Detached Range");

var domTests = [];
var positionTests = [];
var fragTests = [];
var actualIframes = [];
var expectedIframes = [];
for (var i = 0; i < testRanges.length; i++) {
	domTests.push(async_test("Resulting DOM for range " + i + " " + testRanges[i]));
	positionTests.push(async_test("Resulting cursor position for range " + i + " " + testRanges[i]));
	fragTests.push(async_test("Returned fragment for range " + i + " " + testRanges[i]));

	actualIframes.push(document.createElement("iframe"));
	actualIframes[i].style.display = "none";
	document.body.appendChild(actualIframes[i]);

	expectedIframes.push(document.createElement("iframe"));
	expectedIframes[i].style.display = "none";
	document.body.appendChild(expectedIframes[i]);

	testExtractContents(i);
}
</script>
